requirements.txt
fastapi
uvicorn
pandas
matplotlib
seaborn
streamlit
langchain
openai
python-dotenv